[Issue-62] 네비게이션 헤더 - 감정 데이터베이스 초기 설정하는 곳을 감정 커스텀 화면에서 메인 화면으로 변경
